Split up Typescript and Flow generation
