+++ src/typescript.ts
@@ -1,0 +1,63 @@
+import { declarator, isAliasProperty } from './declarator';
+import { EOL, stringifyGenerics, stringifyTypes } from './output';
+
+export default async function typescript() {
+  const { declarations, interfaces } = await declarator();
+
+  let interfacesOutput = '';
+  for (const item of interfaces) {
+    if (interfacesOutput) {
+      interfacesOutput += EOL + EOL;
+    }
+
+    const extendList = item.extends.map(extend => extend.name + stringifyGenerics(extend.generics, true)).join(', ');
+    interfacesOutput += 'export interface ' + item.name + stringifyGenerics(item.generics);
+
+    if (extendList) {
+      interfacesOutput += ` extends ${extendList}`;
+    }
+
+    interfacesOutput += '{' + EOL;
+
+    for (const property of item.properties) {
+      if (property.comment) {
+        interfacesOutput += property.comment + EOL;
+      }
+
+      if (isAliasProperty(property)) {
+        const generics = stringifyGenerics(property.generics, true);
+        interfacesOutput += `${JSON.stringify(property.name)}?: ${
+          item.fallback
+            ? `${property.alias.name + generics} | ${property.alias.name + generics}[];`
+            : `${property.alias.name + generics};`
+        }`;
+      } else {
+        const value = stringifyTypes(property.type);
+        interfacesOutput += `${JSON.stringify(property.name)}?: ${
+          item.fallback ? `${value} | ${value}[];` : `${value};`
+        }`;
+      }
+
+      interfacesOutput += EOL;
+    }
+
+    interfacesOutput += '}';
+  }
+
+  let declarationsOutput = '';
+  for (const declaration of declarations.values()) {
+    if (declarationsOutput) {
+      declarationsOutput += EOL + EOL;
+    }
+
+    if (declaration.export) {
+      declarationsOutput += 'export ';
+    }
+
+    declarationsOutput += `type ${declaration.name + stringifyGenerics(declaration.generics, true)} = ${stringifyTypes(
+      declaration.types,
+    ) + EOL}`;
+  }
+
+  return interfacesOutput + EOL + EOL + declarationsOutput;
+}
